refactor(egopulse): split turn.rs into submodules (guards, compaction, formatting)
