+++ egopulse/src/agent_loop/turn.rs
@@ -4,6 +4,12 @@
 //! 1 本の turn loop としてまとめて扱う。
 
 use crate::agent_loop::SurfaceContext;
+use crate::agent_loop::compaction::maybe_compact_messages;
+use crate::agent_loop::formatting::{
+    format_tool_result, preview_text, sanitize_assistant_response_text,
+    strip_thinking, summarize_tool_calls_with_content, tool_message_content,
+};
+use crate::agent_loop::guards::{is_declarative_only_reply, runtime_guard_messages};
 use crate::agent_loop::session::{
     load_messages_for_turn, persist_phase, persist_phase_once, resolve_chat_id,
 };
@@ -13,12 +19,9 @@ use crate::runtime::{AppState, build_app_state};
 use crate::storage::{StoredMessage, ToolCall as StoredToolCall, call_blocking};
 use crate::tools::ToolExecutionContext;
 use crate::web::sse::AgentEvent;
-use tracing::{info, warn};
+use tracing::warn;
 
 const MAX_TOOL_ITERATIONS: usize = 50;
-const MAX_TOOL_RESULT_CHARS: usize = 16_000;
-const MAX_COMPACTION_SUMMARY_CHARS: usize = 20_000;
-const MAX_TOOL_RESULT_TEXT_CHARS: usize = 200;
 
 /// Sends a one-shot prompt within a named persistent session.
 pub async fn ask_in_session(
@@ -449,103 +452,6 @@ async fn update_tool_call_output(
     .map_err(EgoPulseError::from)
 }
 
-fn format_tool_result(tool_call: &ToolCall, result: &crate::tools::ToolResult) -> String {
-    let mut content = result.content.clone();
-    let details = result.details.clone();
-
-    loop {
-        let mut payload = serde_json::json!({
-            "tool": tool_call.name,
-            "status": if result.is_error { "error" } else { "success" },
-            "result": content,
-        });
-        if let Some(ref d) = details {
-            payload["details"] = d.clone();
-        }
-
-        let serialized = payload.to_string();
-        let char_count = serialized.chars().count();
-
-        if char_count <= MAX_TOOL_RESULT_CHARS {
-            return serialized;
-        }
-
-        // If over limit, first try removing details
-        if details.is_some() {
-            let payload_no_details = serde_json::json!({
-                "tool": tool_call.name,
-                "status": if result.is_error { "error" } else { "success" },
-                "result": content,
-            });
-            let no_details_str = payload_no_details.to_string();
-            if no_details_str.chars().count() <= MAX_TOOL_RESULT_CHARS {
-                return no_details_str;
-            }
-        }
-
-        // Still over limit, truncate content further
-        // Calculate how much we need to reduce content by
-        let excess = char_count.saturating_sub(MAX_TOOL_RESULT_CHARS);
-        let current_content_len = content.chars().count();
-        // Reduce content by excess + buffer for JSON overhead
-        let new_len = current_content_len.saturating_sub(excess + 100);
-        if new_len == 0 {
-            // Can't truncate further, return minimal payload
-            return serde_json::json!({
-                "tool": tool_call.name,
-                "status": if result.is_error { "error" } else { "success" },
-                "result": "...",
-            })
-            .to_string();
-        }
-        content = format!("{}...", content.chars().take(new_len).collect::<String>());
-    }
-}
-
-fn tool_message_content(
-    payload: &str,
-    result: &crate::tools::ToolResult,
-) -> crate::llm::MessageContent {
-    match &result.llm_content {
-        crate::llm::MessageContent::Text(_) => crate::llm::MessageContent::text(payload),
-        crate::llm::MessageContent::Parts(parts) => {
-            let mut content = Vec::with_capacity(parts.len() + 1);
-            content.push(crate::llm::MessageContentPart::InputText {
-                text: payload.to_string(),
-            });
-            content.extend(parts.iter().filter_map(|part| match part {
-                crate::llm::MessageContentPart::InputText { .. } => None,
-                crate::llm::MessageContentPart::InputImage { image_url, detail } => {
-                    Some(crate::llm::MessageContentPart::InputImage {
-                        image_url: image_url.clone(),
-                        detail: detail.clone(),
-                    })
-                }
-            }));
-            crate::llm::MessageContent::parts(content)
-        }
-    }
-}
-
-fn summarize_tool_calls_with_content(content: &str, tool_calls: &[ToolCall]) -> String {
-    let names = tool_calls
-        .iter()
-        .map(|tool_call| tool_call.name.as_str())
-        .collect::<Vec<_>>();
-    if content.trim().is_empty() {
-        format!("[tool_call] {}", names.join(", "))
-    } else {
-        format!("{} [tool_call] {}", content.trim(), names.join(", "))
-    }
-}
-
-fn preview_text(value: &str, max_chars: usize) -> String {
-    if value.chars().count() <= max_chars {
-        return value.to_string();
-    }
-    format!("{}...", value.chars().take(max_chars).collect::<String>())
-}
-
 async fn persist_user_turn_with_compaction(
     state: &AppState,
     context: &SurfaceContext,
@@ -593,446 +499,6 @@ async fn persist_user_turn_with_compaction(
     ))
 }
 
-async fn maybe_compact_messages(
-    state: &AppState,
-    context: &SurfaceContext,
-    chat_id: i64,
-    messages: &[Message],
-    llm: &std::sync::Arc<dyn crate::llm::LlmProvider>,
-) -> Result<Vec<Message>, EgoPulseError> {
-    if messages.len() <= state.config.max_session_messages {
-        return Ok(messages.to_vec());
-    }
-
-    archive_conversation(&state.config.data_dir, &context.channel, chat_id, messages).await;
-
-    let keep_recent = state.config.compact_keep_recent.min(messages.len());
-    if keep_recent == messages.len() {
-        return Ok(messages.to_vec());
-    }
-
-    let split_at = messages.len() - keep_recent;
-    let old_messages = &messages[..split_at];
-    let recent_messages = &messages[split_at..];
-
-    let mut summary_input = String::new();
-    for message in old_messages {
-        let role = &message.role;
-        let text = message_to_text(message);
-        summary_input.push_str(&format!("[{role}]: {text}\n\n"));
-    }
-    summary_input = truncate_compaction_summary_input(summary_input);
-
-    let summarize_prompt = "Summarize the following conversation concisely, preserving key facts, decisions, tool results, and context needed to continue the conversation. Be brief but thorough.";
-    let summarize_messages = vec![Message::text(
-        "user",
-        format!("{summarize_prompt}\n\n---\n\n{summary_input}"),
-    )];
-    let timeout_secs = state.config.compaction_timeout_secs;
-    let summary_result = tokio::time::timeout(
-        std::time::Duration::from_secs(timeout_secs),
-        llm.send_message("You are a helpful summarizer.", summarize_messages, None),
-    )
-    .await;
-
-    let summary = match summary_result {
-        Ok(Ok(response)) => strip_thinking(&response.content),
-        Ok(Err(error)) => {
-            warn!("compaction summarization failed: {error}; falling back to recent messages");
-            return Ok(recent_messages.to_vec());
-        }
-        Err(_) => {
-            warn!(
-                "compaction summarization timed out after {timeout_secs}s for {}:{}; falling back to recent messages",
-                context.channel, chat_id
-            );
-            return Ok(recent_messages.to_vec());
-        }
-    };
-    if summary.trim().is_empty() {
-        warn!("compaction summarization returned empty text; falling back to recent messages");
-        return Ok(recent_messages.to_vec());
-    }
-
-    let mut compacted = vec![Message::text(
-        "user",
-        format!("[Conversation Summary]\n{summary}"),
-    )];
-    if !matches!(recent_messages.first(), Some(message) if message.role == "assistant") {
-        compacted.push(Message::text(
-            "assistant",
-            "Understood, I have the conversation context. How can I help?",
-        ));
-    }
-
-    for message in recent_messages {
-        append_compacted_message(&mut compacted, message);
-    }
-
-    if matches!(compacted.last(), Some(last) if last.role == "assistant") {
-        compacted.pop();
-    }
-
-    Ok(compacted)
-}
-
-async fn archive_conversation(data_dir: &str, channel: &str, chat_id: i64, messages: &[Message]) {
-    let data_dir = data_dir.to_string();
-    let channel = channel.to_string();
-    let messages = messages.to_vec();
-    let join_channel = channel.clone();
-    let join_result = tokio::task::spawn_blocking(move || {
-        archive_conversation_blocking(&data_dir, &channel, chat_id, &messages);
-    })
-    .await;
-
-    if let Err(error) = join_result {
-        warn!(
-            "failed to join archive task for {}:{}: {error}",
-            join_channel, chat_id
-        );
-    }
-}
-
-fn archive_conversation_blocking(
-    data_dir: &str,
-    channel: &str,
-    chat_id: i64,
-    messages: &[Message],
-) {
-    let now = chrono::Utc::now().format("%Y%m%d-%H%M%S");
-    let unique_suffix = uuid::Uuid::new_v4().simple();
-    let channel_dir = if channel.trim().is_empty() {
-        "unknown"
-    } else {
-        channel.trim()
-    };
-    let dir = std::path::PathBuf::from(data_dir)
-        .join("groups")
-        .join(channel_dir)
-        .join(chat_id.to_string())
-        .join("conversations");
-
-    if let Err(error) = std::fs::create_dir_all(&dir) {
-        warn!("failed to create archive dir {}: {error}", dir.display());
-        return;
-    }
-
-    let path = dir.join(format!("{now}-{unique_suffix}.md"));
-    let mut content = String::new();
-    for message in messages {
-        let role = &message.role;
-        let text = message_to_archive_text(message);
-        content.push_str(&format!("## {role}\n\n{text}\n\n---\n\n"));
-    }
-
-    if let Err(error) = std::fs::write(&path, content) {
-        warn!(
-            "failed to archive conversation to {}: {error}",
-            path.display()
-        );
-    } else {
-        info!(
-            "archived conversation ({} messages) to {}",
-            messages.len(),
-            path.display()
-        );
-    }
-}
-
-/// Truncate the compaction summary input by character count, not by bytes.
-///
-/// The limit keeps UTF-8 text intact and appends `\n... (truncated)` when the
-/// input exceeds `MAX_COMPACTION_SUMMARY_CHARS` characters.
-fn truncate_compaction_summary_input(mut summary_input: String) -> String {
-    if summary_input.chars().count() <= MAX_COMPACTION_SUMMARY_CHARS {
-        return summary_input;
-    }
-
-    let cutoff = summary_input
-        .char_indices()
-        .nth(MAX_COMPACTION_SUMMARY_CHARS)
-        .map(|(idx, _)| idx)
-        .unwrap_or(summary_input.len());
-    summary_input.truncate(cutoff);
-    summary_input.push_str("\n... (truncated)");
-    summary_input
-}
-
-fn append_compacted_message(compacted: &mut Vec<Message>, message: &Message) {
-    let Some(last) = compacted.last_mut() else {
-        compacted.push(message.clone());
-        return;
-    };
-
-    if can_merge_compacted_messages(last, message) {
-        let merged = format!(
-            "{}\n{}",
-            last.content.as_text_lossy(),
-            message.content.as_text_lossy()
-        );
-        last.content = crate::llm::MessageContent::text(merged);
-        return;
-    }
-
-    compacted.push(message.clone());
-}
-
-fn can_merge_compacted_messages(left: &Message, right: &Message) -> bool {
-    left.role == right.role
-        && left.tool_calls.is_empty()
-        && right.tool_calls.is_empty()
-        && left.tool_call_id.is_none()
-        && right.tool_call_id.is_none()
-        && matches!(left.content, crate::llm::MessageContent::Text(_))
-        && matches!(right.content, crate::llm::MessageContent::Text(_))
-}
-
-fn message_to_text(message: &Message) -> String {
-    let should_strip_thinking = should_strip_thinking_for_role(&message.role);
-    let content = match &message.content {
-        crate::llm::MessageContent::Text(text) => {
-            if should_strip_thinking {
-                strip_thinking(text)
-            } else {
-                text.clone()
-            }
-        }
-        crate::llm::MessageContent::Parts(parts) => parts
-            .iter()
-            .map(|part| match part {
-                crate::llm::MessageContentPart::InputText { text } => {
-                    if should_strip_thinking {
-                        strip_thinking(text)
-                    } else {
-                        text.clone()
-                    }
-                }
-                crate::llm::MessageContentPart::InputImage { .. } => "[image]".to_string(),
-            })
-            .collect::<Vec<_>>()
-            .join("\n"),
-    };
-    if message.tool_call_id.is_some() {
-        let payload = tool_result_payload(message).unwrap_or_default();
-        let body = strip_thinking(&tool_result_body(payload));
-        let truncated = truncate_summary_text(&body, MAX_TOOL_RESULT_TEXT_CHARS);
-        let prefix = if is_tool_error_message(message) {
-            "[tool_error]: "
-        } else {
-            "[tool_result]: "
-        };
-        return format!("{prefix}{truncated}");
-    }
-
-    let mut parts = Vec::new();
-    if !content.trim().is_empty() {
-        parts.push(content);
-    }
-    for tool_call in &message.tool_calls {
-        parts.push(format!(
-            "[tool_use: {}({})]",
-            tool_call.name, tool_call.arguments
-        ));
-    }
-    parts.join("\n")
-}
-
-fn is_tool_error_message(message: &Message) -> bool {
-    let Some(payload) = tool_result_payload(message) else {
-        return false;
-    };
-    serde_json::from_str::<serde_json::Value>(payload)
-        .ok()
-        .and_then(|value| {
-            value
-                .get("status")
-                .and_then(|status| status.as_str())
-                .map(|status| status == "error")
-        })
-        .unwrap_or(false)
-}
-
-fn tool_result_body(payload: &str) -> String {
-    match serde_json::from_str::<serde_json::Value>(payload) {
-        Ok(value) => value
-            .get("result")
-            .and_then(|result| result.as_str())
-            .map(ToString::to_string)
-            .unwrap_or_else(|| payload.to_string()),
-        Err(_) => payload.to_string(),
-    }
-}
-
-fn sanitize_assistant_response_text(content: &str) -> String {
-    strip_thinking(content.trim())
-}
-
-fn should_strip_thinking_for_role(role: &str) -> bool {
-    matches!(role, "assistant" | "tool")
-}
-
-fn message_to_archive_text(message: &Message) -> String {
-    let should_strip_thinking = should_strip_thinking_for_role(&message.role);
-    let content = match &message.content {
-        crate::llm::MessageContent::Text(text) => {
-            if should_strip_thinking {
-                strip_thinking(text)
-            } else {
-                text.clone()
-            }
-        }
-        crate::llm::MessageContent::Parts(parts) => parts
-            .iter()
-            .map(|part| match part {
-                crate::llm::MessageContentPart::InputText { text } => {
-                    if should_strip_thinking {
-                        strip_thinking(text)
-                    } else {
-                        text.clone()
-                    }
-                }
-                crate::llm::MessageContentPart::InputImage { image_url, detail } => match detail {
-                    Some(detail) => format!("[image: {image_url} detail={detail}]"),
-                    None => format!("[image: {image_url}]"),
-                },
-            })
-            .collect::<Vec<_>>()
-            .join("\n"),
-    };
-
-    if message.tool_call_id.is_some() {
-        let payload = tool_result_payload(message).unwrap_or_default();
-        let body = strip_thinking(payload);
-        let prefix = if is_tool_error_message(message) {
-            "[tool_error]: "
-        } else {
-            "[tool_result]: "
-        };
-        return format!("{prefix}{body}");
-    }
-
-    let mut parts = Vec::new();
-    if !content.trim().is_empty() {
-        parts.push(content);
-    }
-    for tool_call in &message.tool_calls {
-        parts.push(format!(
-            "[tool_use: {}({})]",
-            tool_call.name, tool_call.arguments
-        ));
-    }
-    parts.join("\n")
-}
-
-fn tool_result_payload(message: &Message) -> Option<&str> {
-    match &message.content {
-        crate::llm::MessageContent::Text(text) => Some(text.as_str()),
-        crate::llm::MessageContent::Parts(parts) => parts.iter().find_map(|part| match part {
-            crate::llm::MessageContentPart::InputText { text } => Some(text.as_str()),
-            crate::llm::MessageContentPart::InputImage { .. } => None,
-        }),
-    }
-}
-
-fn truncate_summary_text(text: &str, max_chars: usize) -> String {
-    let char_count = text.chars().count();
-    if char_count <= max_chars {
-        return text.to_string();
-    }
-    let truncated = text.chars().take(max_chars).collect::<String>();
-    format!("{truncated}...")
-}
-
-fn runtime_guard_messages(
-    messages: &[Message],
-    assistant_text: &str,
-    guard_text: &str,
-) -> Vec<Message> {
-    let mut retry_messages = messages.to_vec();
-    retry_messages.push(Message::text("assistant", assistant_text.to_string()));
-    retry_messages.push(Message::text("user", guard_text.to_string()));
-    retry_messages
-}
-
-/// `<think`>`...`</think`>` や `<thought`>`...`</thought`>` などの
-/// thinking タグブロックをモデル出力から除去する。
-/// microclaw agent_engine.rs から移植。
-fn strip_thinking(text: &str) -> String {
-    fn strip_tag_blocks(input: &str, open: &str, close: &str) -> String {
-        let mut result = String::with_capacity(input.len());
-        let mut rest = input;
-        while let Some(start) = rest.find(open) {
-            result.push_str(&rest[..start]);
-            if let Some(end) = rest[start..].find(close) {
-                rest = &rest[start + end + close.len()..];
-            } else {
-                rest = "";
-                break;
-            }
-        }
-        result.push_str(rest);
-        result
-    }
-
-    let no_think = strip_tag_blocks(text, "<think>", "</think>");
-    let no_thought = strip_tag_blocks(&no_think, "<thought>", "</thought>");
-    let no_thinking = strip_tag_blocks(&no_thought, "<thinking>", "</thinking>");
-    let no_reasoning = strip_tag_blocks(&no_thinking, "<reasoning>", "</reasoning>");
-    no_reasoning.trim().to_string()
-}
-
-/// レスポンスが「宣言だけしてツールを実行しない」パターンに一致するか判定する。
-/// microclaw の runtime guard パターンから移植。
-fn is_declarative_only_reply(text: &str) -> bool {
-    let lower = text.to_lowercase();
-    let english_patterns = [
-        "i'll ",
-        "i will ",
-        "i'll go ahead and ",
-        "let me ",
-        "sure, ",
-        "of course, ",
-        "i'd be happy to ",
-        "absolutely, ",
-        "i can help with that",
-        "great, i'll",
-        "alright, i'll",
-        "okay, i'll",
-        "sure thing",
-        "i'm on it",
-    ];
-    let japanese_prefixes = [
-        "了解しました",
-        "承知しました",
-        "わかりました",
-        "かしこまりました",
-        "はい、",
-        "では、",
-        "それでは、",
-        "今から",
-    ];
-    let japanese_action_markers = [
-        "実行します",
-        "確認します",
-        "試してみます",
-        "やってみます",
-        "見てみます",
-        "調べます",
-        "作成します",
-        "書き込みます",
-        "進めます",
-    ];
-    // 短いテキスト（≤200文字）だけを対象とする。長い応答は「宣言のみ」ではない。
-    if lower.trim().chars().count() > 200 {
-        return false;
-    }
-    let trimmed = text.trim();
-    english_patterns.iter().any(|p| lower.trim().starts_with(p))
-        || japanese_prefixes.iter().any(|p| trimmed.starts_with(p))
-        || japanese_action_markers.iter().any(|p| trimmed.contains(p))
-}
-
 #[cfg(test)]
 mod tests {
     use std::sync::Arc;
@@ -1041,10 +507,11 @@ mod tests {
     use secrecy::SecretString;
     use serial_test::serial;
 
-    use crate::agent_loop::turn::{
-        is_declarative_only_reply, message_to_archive_text, message_to_text, strip_thinking,
-        truncate_compaction_summary_input,
+    use crate::agent_loop::compaction::truncate_compaction_summary_input;
+    use crate::agent_loop::formatting::{
+        message_to_archive_text, message_to_text, strip_thinking,
     };
+    use crate::agent_loop::guards::is_declarative_only_reply;
     use crate::agent_loop::{SurfaceContext, process_turn};
     use crate::config::{Config, ProviderConfig};
     use crate::error::{EgoPulseError, LlmError};
